requirements.txt
jinja2
imgkit
discord
chatterbot
pillow